requirements.txt
# Core FastAPI and ASGI server
fastapi==0.122.0
uvicorn==0.38.0
python-multipart==0.0.20
starlette==0.50.0
python-dotenv
# LangChain and LangGraph
langchain-core==1.1.0
langchain-google-genai==3.2.0
langgraph==1.0.4
langgraph-checkpoint==3.0.1
langgraph-checkpoint-postgres==3.0.1
langgraph-prebuilt==1.0.5
langsmith==0.4.48

# Google AI
google-genai==1.52.0
google-ai-generativelanguage==0.9.0
google-api-core==2.28.1
google-auth==2.43.0
googleapis-common-protos==1.72.0

# AWS S3
boto3

# Database
psycopg[binary,pool]

# Environment and utilities
python-dotenv==1.2.1
tenacity==9.1.2
pydantic==2.12.4
pydantic_core==2.41.5

# Core dependencies (auto-installed but pinned for stability)
anyio==4.11.0
certifi==2025.11.12
httpx==0.28.1
idna==3.11
requests==2.32.5
PyYAML==6.0.3
